backend/generate-excel/requirements.txt
openpyxl>=3.1.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
reportlab>=4.0.0